decimal step logic
