+++ tests/index.test.bac.js
@@ -740,7 +740,7 @@ describe.skip('InputNumber', () => {
     it('value decimal over six decimal not be scientific notation', () => {
       class Demo extends React.Component {
         render() {
-          return <InputNumber ref="inputNum" precision={7} step={0.0000001} />;
+          return <InputNumber ref="inputNum" mergedPPrecision={7} step={0.0000001} />;
         }
       }
       example = ReactDOM.render(<Demo />, container);
@@ -1472,7 +1472,7 @@ describe.skip('InputNumber', () => {
           return (
             <div>
               <InputNumber onChange={this.onPrecisionChange} />
-              <InputNumber precision={precision} defaultValue={1.23} />
+              <InputNumber mergedPPrecision={precision} defaultValue={1.23} />
             </div>
           );
         }
@@ -1711,7 +1711,7 @@ describe.skip('InputNumber', () => {
 
         render() {
           return (
-            <InputNumber ref="inputNum" precision={2} defaultValue={2} onChange={this.onChange} />
+            <InputNumber ref="inputNum" mergedPPrecision={2} defaultValue={2} onChange={this.onChange} />
           );
         }
       }
@@ -1745,7 +1745,7 @@ describe.skip('InputNumber', () => {
 
         render() {
           return (
-            <InputNumber ref="inputNum" precision={2} defaultValue={2} onChange={this.onChange} />
+            <InputNumber ref="inputNum" mergedPPrecision={2} defaultValue={2} onChange={this.onChange} />
           );
         }
       }
@@ -2079,7 +2079,7 @@ describe('Mobile inputNumber use TouchEvents', () => {
             onBlur={({ target: { value } }) => {
               expect(value).to.be('1');
             }}
-            precision={0}
+            mergedPPrecision={0}
             ref="inputNum"
           />
         );
